+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/accountSettings/selectAppearance.tsx
@@ -1,23 +1,16 @@
 import React, { FC } from 'react'
 import { View, Text, TouchableOpacity, useTheme } from '@avalabs/k2-alpine'
-import { useDispatch, useSelector } from 'react-redux'
-import {
-  Appearance,
-  selectSelectedAppearance,
-  setSelectedAppearance
-} from 'store/settings/appearance'
+import { useSelector } from 'react-redux'
+import { Appearance, selectSelectedAppearance } from 'store/settings/appearance'
 import { SvgProps } from 'react-native-svg'
+import { useColorScheme } from 'common/contexts/ColorSchemeProvider'
 import SystemAppearanceIcon from '../../../../assets/icons/system_appearance.svg'
 import LightAppearanceIcon from '../../../../assets/icons/light_appearance.svg'
 import DarkAppearanceIcon from '../../../../assets/icons/dark_appearance.svg'
 
 const SelectAppearanceScreen = (): JSX.Element => {
-  const dispatch = useDispatch()
   const selectedAppearance = useSelector(selectSelectedAppearance)
-
-  const handleSelectAppearance = (appearance: Appearance): void => {
-    dispatch(setSelectedAppearance(appearance))
-  }
+  const { toggleColorScheme } = useColorScheme()
 
   return (
     <View
@@ -36,19 +29,19 @@ const SelectAppearanceScreen = (): JSX.Element => {
         <AppearanceComponent
           isSelected={selectedAppearance === Appearance.System}
           appearance="System"
-          onPress={() => handleSelectAppearance(Appearance.System)}
+          onPress={() => toggleColorScheme(Appearance.System)}
           Icon={SystemAppearanceIcon}
         />
         <AppearanceComponent
           isSelected={selectedAppearance === Appearance.Light}
           appearance="Light"
-          onPress={() => handleSelectAppearance(Appearance.Light)}
+          onPress={() => toggleColorScheme(Appearance.Light)}
           Icon={LightAppearanceIcon}
         />
         <AppearanceComponent
           isSelected={selectedAppearance === Appearance.Dark}
           appearance="Dark"
-          onPress={() => handleSelectAppearance(Appearance.Dark)}
+          onPress={() => toggleColorScheme(Appearance.Dark)}
           Icon={DarkAppearanceIcon}
         />
       </View>
